requirements.txt
black==25.1.0
pylint==3.3.6
pytest==8.3.5
python-dotenv==1.0.1
tiktoken==0.9.0
openai==1.65.0